HACK agent: make kata-agent.service depend on virtio-serial
The virtio-serial device may not be initialized when kata-agent is
launched.  This problem is easy to trigger with systemd and a modular
kernel.

Add an explicit dependency on the virtio-serial channel so that
kata-agent startup succeeds.  This is kind of a hack since we don't want
this dependency when using AF_VSOCK :(.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>